fix: stop the aiohttp app when the metricq HistoryClient is stopped and vice versa
We setup a watchdog that (a)waits for the HistoryClient to stop for whatever reason and then emulates the aiohttp signal handler.
If the app is stopped first, this watchdog is cancelled so that we don't get some kind of recursive mess.
Also, the on the app cleanup (caused by the signal handler), we properly stop the HistoryClient.
This does not deal with any open client connection tasks that may or may not need some extra cancellation.